Fix unsigned long <-> unsigned int issue.

+++ lib/gladman-fcrypt/hmac.c
@@ -47,7 +47,7 @@ INTERNAL void hmac_sha1_begin(hmac_ctx cx[1])
 }
 
 /* input the HMAC key (can be called multiple times)    */
-INTERNAL int hmac_sha1_key(const unsigned char key[], unsigned long key_len, hmac_ctx cx[1])
+INTERNAL int hmac_sha1_key(const unsigned char key[], unsigned int key_len, hmac_ctx cx[1])
 {
     if(cx->klen == HMAC_IN_DATA)                /* error if further key input   */
         return HMAC_BAD_MODE;                   /* is attempted in data mode    */
@@ -71,7 +71,7 @@ INTERNAL int hmac_sha1_key(const unsigned char key[], unsigned long key_len, hma
 
 /* input the HMAC data (can be called multiple times) - */
 /* note that this call terminates the key input phase   */
-INTERNAL void hmac_sha1_data(const unsigned char data[], unsigned long data_len, hmac_ctx cx[1])
+INTERNAL void hmac_sha1_data(const unsigned char data[], unsigned int data_len, hmac_ctx cx[1])
 {   unsigned int i;
 
     if(cx->klen != HMAC_IN_DATA)                /* if not yet in data phase */
