+++ lib/gladman-fcrypt/hmac.h
@@ -61,9 +61,9 @@ typedef struct
 #ifdef BUILDING_FCRYPT
 INTERNAL void hmac_sha1_begin(hmac_ctx cx[1]);
 
-INTERNAL int  hmac_sha1_key(const unsigned char key[], unsigned long key_len, hmac_ctx cx[1]);
+INTERNAL int  hmac_sha1_key(const unsigned char key[], unsigned int key_len, hmac_ctx cx[1]);
 
-INTERNAL void hmac_sha1_data(const unsigned char data[], unsigned long data_len, hmac_ctx cx[1]);
+INTERNAL void hmac_sha1_data(const unsigned char data[], unsigned int data_len, hmac_ctx cx[1]);
 
 INTERNAL void hmac_sha1_end(unsigned char mac[], unsigned long mac_len, hmac_ctx cx[1]);
 #endif
